pyproject.toml
[tool.poetry]
name = "supername"
version = "0.1.0"
description = "supername project"
authors = ["trbouma <trbouma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.4"
requests = "^2.31.0"
click = "^8.1.7"
pyyaml = "^6.0.1"
fido2 = "^1.1.3"
webauthn = "^2.1.0"

[tool.poetry.scripts]
supername = "supername.main:cli"
super = "supername.main:cli"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


